Details page and secondary page duplicate removed
