+++ app/release/output.json
@@ -1,1 +1,1 @@
-[{"outputType":{"type":"APK"},"apkInfo":{"type":"MAIN","splits":[],"versionCode":4},"path":"app-release.apk","properties":{"packageId":"com.qatarmuseums.qatarmuseumsapp","split":"","minSdkVersion":"21"}}]
+[{"outputType":{"type":"APK"},"apkInfo":{"type":"MAIN","splits":[],"versionCode":6},"path":"app-release.apk","properties":{"packageId":"com.qatarmuseums.qatarmuseumsapp","split":"","minSdkVersion":"21"}}]
