+++ app/src/main/java/com/qatarmuseums/qatarmuseumsapp/apicall/APIInterface.java
@@ -1,7 +1,7 @@
 package com.qatarmuseums.qatarmuseumsapp.apicall;
 
 
-import com.qatarmuseums.qatarmuseumsapp.commonpage.CommonModel;
+import com.qatarmuseums.qatarmuseumsapp.commonlistpage.CommonListModel;
 import com.qatarmuseums.qatarmuseumsapp.culturepass.LoginData;
 import com.qatarmuseums.qatarmuseumsapp.culturepass.TokenForPushNotification;
 import com.qatarmuseums.qatarmuseumsapp.detailspage.RegistrationDetailsModel;
@@ -42,13 +42,13 @@ public interface APIInterface {
     Call<ArrayList<HomeList>> getMuseumsList(@Path("language") String language);
 
     @GET("{language}/mobile_api/{pageName}")
-    Call<ArrayList<CommonModel>> getCommonPageList(@Path("language") String language,
-                                                   @Path("pageName") String pageName);
+    Call<ArrayList<CommonListModel>> getCommonPageList(@Path("language") String language,
+                                                       @Path("pageName") String pageName);
 
     @GET("{language}/mobile_api/{pageName}")
-    Call<ArrayList<CommonModel>> getCommonPageListWithID(@Path("language") String language,
-                                                         @Path("pageName") String pageName,
-                                                         @Query("museum_id") String museumId);
+    Call<ArrayList<CommonListModel>> getCommonPageListWithID(@Path("language") String language,
+                                                             @Path("pageName") String pageName,
+                                                             @Query("museum_id") String museumId);
 
     @GET("{language}/mobile_api/getpublicartdetail.json")
     Call<ArrayList<PublicArtModel>> getPublicArtsDetails(@Path("language") String language,
@@ -67,8 +67,8 @@ public interface APIInterface {
     Call<ArrayList<ParkList>> getParkDetails(@Path("language") String language);
 
     @GET("{language}/mobile_api/museum_collection_category.json")
-    Call<ArrayList<CommonModel>> getCollectionList(@Path("language") String language,
-                                                   @Query("museum_id") String museumId);
+    Call<ArrayList<CommonListModel>> getCollectionList(@Path("language") String language,
+                                                       @Query("museum_id") String museumId);
 
 
     @GET("{language}/mobile_api/museum-detail.json")
@@ -140,16 +140,16 @@ public interface APIInterface {
                                                                   @Query("nid") String nid);
 
     @GET("{language}/mobile_api/nmoq_list_day.json")
-    Call<ArrayList<CommonModel>> getTourList(@Path("language") String language);
+    Call<ArrayList<CommonListModel>> getTourList(@Path("language") String language);
 
     @GET("{language}/mobile_api/list_facility_category.json")
-    Call<ArrayList<CommonModel>> getFacilityList(@Path("language") String language);
+    Call<ArrayList<CommonListModel>> getFacilityList(@Path("language") String language);
 
     @GET("{language}/mobile_api/nmoq_list_partner.json")
-    Call<ArrayList<CommonModel>> getTravelData(@Path("language") String language);
+    Call<ArrayList<CommonListModel>> getTravelData(@Path("language") String language);
 
     @GET("{language}/mobile_api/nmoq_special_event.json")
-    Call<ArrayList<CommonModel>> getSpecialEvents(@Path("language") String language);
+    Call<ArrayList<CommonListModel>> getSpecialEvents(@Path("language") String language);
 
     @GET("{language}/mobile_api/list_tour_per_day.json")
     Call<ArrayList<TourDetailsModel>> getTourDetails(@Path("language") String language,
