+++ app/src/main/java/com/qatarmuseums/qatarmuseumsapp/commonlistpage/RecyclerTouchListener.java
@@ -1,4 +1,4 @@
-package com.qatarmuseums.qatarmuseumsapp.commonpage;
+package com.qatarmuseums.qatarmuseumsapp.commonlistpage;
 
 
 import android.content.Context;
