+++ app/src/main/java/com/qatarmuseums/qatarmuseumsapp/commonlistsecondary/SecondaryListAdapter.java
@@ -1,9 +1,11 @@
-package com.qatarmuseums.qatarmuseumsapp.toursecondarylist;
+package com.qatarmuseums.qatarmuseumsapp.commonlistsecondary;
 
 
+import android.app.Activity;
 import android.content.Context;
 import android.support.annotation.NonNull;
 import android.support.v7.widget.RecyclerView;
+import android.util.DisplayMetrics;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -12,20 +14,22 @@ import android.widget.RelativeLayout;
 import android.widget.TextView;
 
 import com.qatarmuseums.qatarmuseumsapp.R;
-import com.qatarmuseums.qatarmuseumsapp.commonpage.RecyclerTouchListener;
+import com.qatarmuseums.qatarmuseumsapp.commonlistpage.RecyclerTouchListener;
 import com.qatarmuseums.qatarmuseumsapp.home.GlideApp;
 import com.qatarmuseums.qatarmuseumsapp.tourdetails.TourDetailsModel;
 
 import java.lang.ref.WeakReference;
 import java.util.List;
 
-public class TourSecondaryListAdapter extends RecyclerView.Adapter<TourSecondaryListAdapter.MyViewHolder> {
+public class SecondaryListAdapter extends RecyclerView.Adapter<SecondaryListAdapter.MyViewHolder> {
 
     private final Context mContext;
     private final RecyclerTouchListener.ItemClickListener listener;
     private List<TourDetailsModel> commonModelList;
 
     public class MyViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
+        private final int height;
+        RelativeLayout rowItem;
         TextView name;
         TextView tourDayTxt;
         TextView tourDateTxt;
@@ -45,8 +49,13 @@ public class TourSecondaryListAdapter extends RecyclerView.Adapter<TourSecondary
             tourDayTxt = view.findViewById(R.id.tour_day_text);
             tourDateTxt = view.findViewById(R.id.tour_date_text);
             tourTitleTxt = view.findViewById(R.id.tour_title_text);
+            rowItem = view.findViewById(R.id.row_item);
             name.setAllCaps(false);
             view.setOnClickListener(this);
+            DisplayMetrics displayMetrics = new DisplayMetrics();
+            ((Activity) mContext).getWindowManager().getDefaultDisplay().getMetrics(displayMetrics);
+            height = (int) (displayMetrics.heightPixels * 0.27);
+
         }
 
         @Override
@@ -55,7 +64,7 @@ public class TourSecondaryListAdapter extends RecyclerView.Adapter<TourSecondary
         }
     }
 
-    TourSecondaryListAdapter(Context context, List<TourDetailsModel> commonModelList, RecyclerTouchListener.ItemClickListener listener) {
+    SecondaryListAdapter(Context context, List<TourDetailsModel> commonModelList, RecyclerTouchListener.ItemClickListener listener) {
         this.commonModelList = commonModelList;
         this.mContext = context;
         this.listener = listener;
@@ -65,7 +74,7 @@ public class TourSecondaryListAdapter extends RecyclerView.Adapter<TourSecondary
 
     @NonNull
     @Override
-    public TourSecondaryListAdapter.MyViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
+    public SecondaryListAdapter.MyViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         View itemView = LayoutInflater.from(parent.getContext())
                 .inflate(R.layout.common_list_row, parent, false);
 
@@ -73,9 +82,12 @@ public class TourSecondaryListAdapter extends RecyclerView.Adapter<TourSecondary
     }
 
     @Override
-    public void onBindViewHolder(@NonNull TourSecondaryListAdapter.MyViewHolder holder, int position) {
+    public void onBindViewHolder(@NonNull SecondaryListAdapter.MyViewHolder holder, int position) {
         TourDetailsModel model = commonModelList.get(position);
         holder.name.setText(model.getTourTitle());
+        ViewGroup.LayoutParams params = holder.rowItem.getLayoutParams();
+        params.height = holder.height;
+        holder.rowItem.setLayoutParams(params);
 
         if (model.getTourImage().size() > 0) {
             GlideApp.with(mContext)
