+++ app/src/main/java/com/qatarmuseums/qatarmuseumsapp/commonpagedatabase/ExhibitionListTableEnglish.java
@@ -17,6 +17,8 @@ public class ExhibitionListTableEnglish {
     @ColumnInfo()
     private String exhibition_latest_image;
     @ColumnInfo()
+    private String exhibition_images;
+    @ColumnInfo()
     private String exhibition_start_date;
     @ColumnInfo()
     private String exhibition_end_date;
@@ -43,7 +45,8 @@ public class ExhibitionListTableEnglish {
                                       String exhibition_long_description,
                                       String exhibition_latitude, String exhibition_longitude,
                                       String museum_id, String exhibition_status,
-                                      String exhibition_display_date) {
+                                      String exhibition_display_date,
+                                      String exhibition_images) {
         this.exhibition_id = exhibition_id;
         this.exhibition_name = exhibition_name;
         this.exhibition_latest_image = exhibition_latest_image;
@@ -57,6 +60,7 @@ public class ExhibitionListTableEnglish {
         this.museum_id = museum_id;
         this.exhibition_status = exhibition_status;
         this.exhibition_display_date = exhibition_display_date;
+        this.exhibition_images = exhibition_images;
     }
 
     @NonNull
@@ -82,6 +86,18 @@ public class ExhibitionListTableEnglish {
 
     public void setExhibition_latest_image(String exhibition_latest_image) {
         this.exhibition_latest_image = exhibition_latest_image;
+    }
+
+    public String getExhibition_images() {
+        return exhibition_images;
+    }
+
+    public void setExhibition_images(String exhibition_images) {
+        this.exhibition_images = exhibition_images;
+    }
+
+    public void setExhibition_display_date(String exhibition_display_date) {
+        this.exhibition_display_date = exhibition_display_date;
     }
 
     public String getExhibition_start_date() {
