+++ app/src/main/java/com/qatarmuseums/qatarmuseumsapp/commonpagedatabase/ExhibitionTableDao.java
@@ -54,14 +54,17 @@ public interface ExhibitionTableDao {
     @Query("UPDATE exhibitionlistenglish SET exhibition_start_date = :startDateFromApi," +
             "exhibition_end_date = :endDateFromApi," +
             "exhibition_latest_image = :imageFromApi," +
+            "exhibition_images = :imagesFromApi," +
             "exhibition_long_description=:descriptionFromApi," +
             "exhibition_latitude=:latitudeFromApi," +
             "exhibition_longitude=:longitudeFromApi," +
             "exhibition_short_description=:shortDescriptionFromApi WHERE exhibition_id=:id")
     void updateExhibitionDetailEnglish(String startDateFromApi, String endDateFromApi,
                                        String imageFromApi, String descriptionFromApi,
-                                       String shortDescriptionFromApi, String latitudeFromApi, String longitudeFromApi,
+                                       String shortDescriptionFromApi, String latitudeFromApi,
+                                       String longitudeFromApi, String imagesFromApi,
                                        String id);
+
     @Query("UPDATE exhibitionlistarabic SET exhibition_start_date = :startDateFromApi," +
             "exhibition_end_date = :endDateFromApi,exhibition_location = :locationFromApi," +
             "museum_id = :museumIdFromApi WHERE exhibition_id=:id")
@@ -71,13 +74,15 @@ public interface ExhibitionTableDao {
     @Query("UPDATE exhibitionlistarabic SET exhibition_start_date = :startDateFromApi," +
             "exhibition_end_date = :endDateFromApi," +
             "exhibition_latest_image = :imageFromApi," +
+            "exhibition_images = :imagesFromApi," +
             "exhibition_long_description=:descriptionFromApi," +
             "exhibition_latitude=:latitudeFromApi," +
             "exhibition_longitude=:longitudeFromApi," +
             "exhibition_short_description=:shortDescriptionFromApi WHERE exhibition_id=:id")
     void updateExhibitionDetailArabic(String startDateFromApi, String endDateFromApi,
                                       String imageFromApi, String descriptionFromApi,
-                                      String shortDescriptionFromApi, String latitudeFromApi, String longitudeFromApi,
+                                      String shortDescriptionFromApi, String latitudeFromApi,
+                                      String longitudeFromApi, String imagesFromApi,
                                       String id);
 
     @Insert
