+++ app/src/main/java/com/qatarmuseums/qatarmuseumsapp/commonpagedatabase/PublicArtsTableArabic.java
@@ -16,27 +16,31 @@ public class PublicArtsTableArabic {
     @PrimaryKey(autoGenerate = false)
     private long public_arts_id;
     @ColumnInfo()
+    private String sort_id;
+    @ColumnInfo()
     private String public_arts_name;
     @ColumnInfo()
     private String public_arts_image;
+    @ColumnInfo()
+    private String public_arts_images;
     @ColumnInfo()
     private String longitude;
     @ColumnInfo()
     private String latitude;
     @ColumnInfo()
     private String description;
-
     @ColumnInfo()
     private String short_description;
 
     public PublicArtsTableArabic(@NonNull long public_arts_id, String public_arts_name,
-                                  String public_arts_image,
-                                  String longitude, String latitude) {
+                                 String public_arts_image,
+                                 String longitude, String latitude, String sort_id) {
         this.public_arts_id = public_arts_id;
         this.public_arts_name = public_arts_name;
-        this.public_arts_image=public_arts_image;
+        this.public_arts_image = public_arts_image;
         this.longitude = longitude;
         this.latitude = latitude;
+        this.sort_id = sort_id;
     }
 
 
@@ -47,6 +51,14 @@ public class PublicArtsTableArabic {
 
     public void setPublic_arts_id(@NonNull long public_arts_id) {
         this.public_arts_id = public_arts_id;
+    }
+
+    public String getSort_id() {
+        return sort_id;
+    }
+
+    public void setSort_id(String sort_id) {
+        this.sort_id = sort_id;
     }
 
     public String getPublic_arts_name() {
@@ -63,6 +75,14 @@ public class PublicArtsTableArabic {
 
     public void setPublic_arts_image(String public_arts_image) {
         this.public_arts_image = public_arts_image;
+    }
+
+    public String getPublic_arts_images() {
+        return public_arts_images;
+    }
+
+    public void setPublic_arts_images(String public_arts_images) {
+        this.public_arts_images = public_arts_images;
     }
 
     public String getLongitude() {
