+++ app/src/main/java/com/qatarmuseums/qatarmuseumsapp/settings/SettingsActivity.java
@@ -21,7 +21,7 @@ import android.widget.TextView;
 
 import com.qatarmuseums.qatarmuseumsapp.LocaleManager;
 import com.qatarmuseums.qatarmuseumsapp.R;
-import com.qatarmuseums.qatarmuseumsapp.commonpage.RecyclerTouchListener;
+import com.qatarmuseums.qatarmuseumsapp.commonlistpage.RecyclerTouchListener;
 import com.qatarmuseums.qatarmuseumsapp.home.HomeActivity;
 
 import java.util.ArrayList;
