+++ app/src/main/res/layout/activity_common_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.qatarmuseums.qatarmuseumsapp.commonpage.CommonActivity">
+    tools:context="com.qatarmuseums.qatarmuseumsapp.commonlistpage.CommonListActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/app_bar"
