+++ app/src/main/res/layout/activity_common_list_secondary.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.qatarmuseums.qatarmuseumsapp.toursecondarylist.TourSecondaryListActivity">
+    tools:context="com.qatarmuseums.qatarmuseumsapp.commonlistsecondary.SecondaryListActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/app_bar"
@@ -190,7 +190,7 @@
             </LinearLayout>
 
             <android.support.v7.widget.RecyclerView
-                android:id="@+id/tour_recycler_view"
+                android:id="@+id/secondary_list_recycler_view"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:background="@color/colorRecyclerBackground"
