+++ app/src/main/res/layout/dining_details.xml
@@ -37,13 +37,13 @@
                 android:background="@drawable/round_corner_white_details_page"
                 android:gravity="center_horizontal"
                 android:orientation="vertical"
-                android:padding="@dimen/padding_25"
                 android:visibility="visible">
 
                 <TextView
                     android:id="@+id/main_title"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
+                    android:layout_marginTop="@dimen/padding_25"
                     android:fontFamily="@font/din_regular"
                     android:textAlignment="center"
                     android:textAllCaps="true"
@@ -63,6 +63,8 @@
                     android:layout_height="wrap_content"
                     android:layout_margin="@dimen/text_margin"
                     android:fontFamily="@font/din_next_lt_pro_regular"
+                    android:paddingEnd="@dimen/padding_25"
+                    android:paddingStart="@dimen/padding_25"
                     android:textAlignment="center"
                     android:textColor="@color/colorBlack"
                     android:textSize="@dimen/font_size_15" />
@@ -73,6 +75,8 @@
                     android:layout_height="wrap_content"
                     android:layout_margin="@dimen/text_margin"
                     android:fontFamily="@font/din_next_lt_pro_regular"
+                    android:paddingEnd="@dimen/padding_25"
+                    android:paddingStart="@dimen/padding_25"
                     android:textAlignment="center"
                     android:textColor="@color/colorBlack"
                     android:textSize="@dimen/font_size_15"
@@ -112,6 +116,8 @@
                         android:layout_marginEnd="@dimen/text_margin"
                         android:layout_marginStart="@dimen/text_margin"
                         android:fontFamily="@font/din_next_lt_pro_regular"
+                        android:paddingEnd="@dimen/padding_25"
+                        android:paddingStart="@dimen/padding_25"
                         android:textAlignment="center"
                         android:textColor="@color/colorBlack"
                         android:textSize="@dimen/font_size_15" />
